refactor: add context in folder home
